add single pipeline endpoint and moving some logic around
