+++ lib/jenkins_pipeline/app.rb
@@ -14,56 +14,37 @@ module JenkinsPipeline
       pipelines.map(&:serialize).to_json
     end
 
-    def pipelines
-      client = JenkinsClient.new
-
-      Dir["config/*.yml"].map do |file|
-        pipeline = YAML.load_file(file)
-        jobs_in_folder = client.all_jobs_from(pipeline)
-        create_pipeline(jobs_in_folder, pipeline)
-      end
-    end
-
-    def create_pipeline(jobs_in_folder, pipeline)
-      jobs = []
-      pipeline_instance = Pipeline.new pipeline["name"]
-      has_incomplete = false
-
-      pipeline["jobs"].each_with_index do |job, index|
-        ran = true
-        result = ""
-        job_in_folder = jobs_in_folder["jobs"].select {|job_in_folder| job["ci_name"] == job_in_folder["name"] }.first
-
-        upstream_job = jobs.select {|j| j.name == job["upstream"]}.first
-        upstream_in_folder = job_in_folder["upstreamProjects"].select {|upstream| upstream["name"] == job["upstream"]}.first
-
-        if (not upstream_job.nil?)
-          current_upstream_build = upstream_job.number
-          if (not upstream_in_folder.nil?)
-            current_upstream_build = upstream_job.number
-            upstream_in_folder_build = upstream_in_folder["nextBuildNumber"] - 1
-
-            if (current_upstream_build != upstream_in_folder_build || has_incomplete == true)
-              result = "danger"
-              ran = false
-              has_incomplete = true
-            end
-          end
-        end
-
-        job_instance = Job.new(job_in_folder, job["name"], result, ran)
-
-        pipeline_instance.add_job job_instance, is_first?(index)
-
-        jobs.push(job_instance)
-      end
-      pipeline_instance
+    get '/api/pipelines/:name' do
+      content_type :json
+      file = pipeline_files.select { |file| file.fetch("url") == params[:name] }.first
+      halt 404 if file.nil?
+      pipeline(file).serialize.to_json
     end
 
     private
 
-    def is_first? index
-      index == 0
+    def pipelines
+      client = JenkinsClient.new
+      pipeline_builder = PipelineBuilder.new
+
+      pipeline_files.map do |pipeline|
+        jobs_in_folder = client.all_jobs_from(pipeline)
+        pipeline_builder.build(jobs_in_folder, pipeline)
+      end
+    end
+
+    def pipeline file
+      client = JenkinsClient.new
+      pipeline_builder = PipelineBuilder.new
+
+      jobs_in_folder = client.all_jobs_from(file)
+      pipeline_builder.build(jobs_in_folder, file)
+    end
+
+    def pipeline_files
+      Dir["config/*.yml"].map do |file|
+        pipeline = YAML.load_file(file)
+      end
     end
   end
 end
